Avoid unnecessary stop in edge case, improve warnings
mkinfit: Do not error out in cases where the fit converges, but the
Jacobian for the untransformed model cost can not be estimated. Give a
warning instead and return NA for the t-test results.

summary.mkinfit: Give a warning message when the covariance matrix can
not be obtained.

A test has been added to containing such an edge case to check that
the warnings are correctly issued and the fit does not terminate.