Improved way to have persistent DLLs for mkinmod
Depends on inline >= 0.16.2 (including the bug fixes from
eddelbuettel/inline#18), which provides 'moveDLL' to store the DLL for a
compiled function in a safe place in case the argument 'dll_dir' is
specified in the call to 'mkinmod'.

Huge thanks to Dirk @eddelbuettel for his review and support
for the work on the inline package.