Address issue revealed by additional tests on CRAN
Remove tests relying on non-convergence of the FOMC fit to the FOCUS A
dataset, as this is platform dependent. After the upgrade, the fit
converges on this system as well, although neither ATLAS is used, nor R
was configured disabling long doubles (these were the conditions under
which the tests failed on CRAN).