Correction of the dimethenamid dataset
The data given in previous mkin versions for "Borstel 2" (Staudenmaier
2013) are in fact a subset of the "Borstel 1" data, with the "Borstel 2"
data simply being further analysed for isomer composition by
Staudenmaier (2013).

While correcting this, the URL of the RAR published by EFSA was
updated, as the EFSA Register of Questions was replaced by the
Open EFSA portal.

This also affects the recent publication in "Environments" [1], where
these two datasets were pooled. However, the influence on the final
results can be expected to be minimal. Out of the 16 observations for
the Borstel soil, five were erroneously duplicated which just changes
the weight of the different sampling points and modifies the variance.
The conclusions of the publication are qualitative and are not
considered to be affected.

[1] https://doi.org/10.3390/environments8080071