Update the dimethenamid vignette with current saemix
Convergence is faster with this version (@ecomets mentioned that there
was a bugfix lately that could lead to faster convergence). However,
if I use too many iterations (i.e. 10 000 as in the last version of
this vignette), I get an error in solving omega.teta during later
iterations, apparently due to overparameterisation of the DFOP model
in this case.