Simplify SFORB analytical solution, whitespace
I do not know why the formulae for b1 and b2 on page 64 of FOCUS
kinetics (2014) were not simplified. Clearly, the term

  k12 * k21 - (k12 + k1output) * k21)

can be simplified to

  - k1output * k21

The test for equivalence of DFOP and SFORB fits verifies that the change
is OK. I also removed trailing whitespaces, substituted tab characters
by two whitespaces and removed indenting of text in paragraphs
describing parameters in roxygen comments to unify formatting.