Make using predefined symbols optional
I got crashes under some circumstances when using symbols stored in the
mkinmod object. One such circumstance was the use of a moved DLL in
combination with parallel computing on a cluster. As I cannot exactly
define at the moment when this happens, it is better to make this
an opt in.