Fix and rebuild documentation, see NEWS
I had to fix the two pathway vignettes, as they did not work with
the released version any more. So they and the multistart vignette
which got some small fixes as well were rebuilt.

Complete rebuild of the online docs with the released version. The
documentation of the 'hierarchial_kinetics' format had to be fixed
as well.