Facilitate centering covariates
The motivation for this is to obtain confidence intervals for
covariate dependent parameters that are valid for a median/mean
value of the covariate.

Implementation by adding an argument to the 'saem' function, and
adapting the relevant functions working with 'saem' objects.

Vignettes, the template and tests were updated to use the new functionality.